requirements.txt
langchain
wikipedia
langchain_community
langchain_mistralai
python-dotenv
pydantic
duckduckgo-search
